requirements.txt
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
PyYAML==5.1.1
protobuf==3.8.0
tensorboardX==1.7
torch==1.1.0
torchvision==0.3.0
virtualenv==16.6.0
